requirements.txt
python-socketio==5.8.0
eventlet==0.33.3
pyaudio==0.2.11
pydub==0.25.1
mutagen==1.47.0
Pillow==10.0.0
pygame==2.5.2
ffmpeg-python==0.2.0 